+++ CHANGES.md
@@ -6,6 +6,22 @@ Version 0.2.0
 
 To be released.
 
+ -  Added stops (periods/commas/interpuncts) normalizer.
+
+    Haskell API-wise, the below types and functions were added:
+
+     -  `Text.Seonbi.Punctuation.Stops` data type
+     -  `Text.Seonbi.Punctuation.normalizeStops` function
+     -  `Text.Seonbi.Punctuation.horizontalStops` function
+     -  `Text.Seonbi.Punctuation.horizontalStopsWithSlashes` function
+     -  `Text.Seonbi.Punctuation.verticalStops` function
+     -  `Text.Seonbi.Facade.StopOption` data type
+     -  `stop` field in `Text.Seonbi.Facade.Configuration` data constructor
+
+    CLI-wise, the `-s`/`--stop` option was added.
+
+    HTTP API-wise, the optional field `"stop"` was added.
+
  -  Updated the *Standard Korean Language Dictionary* data
     (*data/ko-kr-stdict.tsv*) to the revision 2020-03-25.